Add Docstring ABC to common
